Support freedesktop.org Idle Inhibition Service
This will only activate if python3-dbus-next is installed.
It can be disabled with --noinhibit on the command line or inhibit=false in the
config file.  These options exist only as a temporary measure in case the new
dbus support causes problems for some users.  They will most likely be removed
in a future version.